# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: |
    Internal static website hosting. Every site is served at the root of its
    own host, `<sitename>.<owner>.<base>/` (for example
    `todo.alice.example.com/`). The owner host `<owner>.<base>/` serves the
    person's or team's index page. Until the owner's certificate is ready it
    also serves their sites at `<owner>.<base>/<sitename>/`; after that,
    those addresses, and `/api/sites/<sitename>/...` on the owner host,
    redirect to each site's own host with path and query kept (`301` for GET
    and HEAD, `308` otherwise). v1.2 addresses of the form
    `<owner>--<sitename>.<base>` only redirect. Always use the `url` from the
    latest response rather than composing an address. Uploads may
    be gzip-compressed tar or ZIP archives; nested regular-file downloads such
    as ZIP and DMG artifacts are supported.

    A site is owned by a person or a team. Team members change a team's
    sites through owner-qualified collaboration routes with optimistic
    `ETag` / `If-Match` protection.

    Every site has an access level: `only_me` (the owner, or the team's
    members; the default for a new site), `specific` (plus named viewers),
    `company` (anyone signed in with the link),
    `listed` (company, plus showcase and search), or `network` (anyone who
    can reach the server, no sign-in). `network` is requested with a reason
...
    network site's pages, assets and state, and change nothing.

...

...

...

...
servers:
...
paths:
...
  /api/sites:
    get:
      summary: List sites
      description: Returns the authenticated user's sites with daily server-side analytics. Admins see all sites.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
      responses:
        '200':
          description: List of sites
          headers:
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'
        '400':
          $ref: '#/components/responses/SkillVersionRequired'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/users/{username}/export:
    get:
      summary: Export everything held about a disabled person
      description: Admin only, browser session only. Streams one zip - `account.json`, `teams.json`, `viewer-grants.json`, `api-keys.json`, `connected-apps.json`, `sessions.json` (metadata only, never a key, token or secret), `pending-grants.json` (viewer and team grants waiting for their provider-vouched email), `visits.jsonl` (the access-log rows of their own visits), `sites/<name>/` for every site they own including recently deleted ones (live files under `files/`, `site.json`, `saved-data.json`, `saved-data-history.json`, `versions.json`, `assets.json`, asset bytes under `assets/<id>`), and `audit-events.jsonl` (every audited action they took). Audited as `admin_user_export`.
      tags: [Admin]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: The name is a team
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The person can still sign in; disable them first
  /api/admin/users/{username}/erase:
    post:
      summary: Delete a disabled person and all their data
      description: Admin only, browser session only. Permanent, with no recovery window - their sites (live and recently deleted) with files, saved data and its history, versions and assets (bucket objects queued for the retire sweep), API keys, connected apps, sessions, team memberships, viewer grants they hold, pending grants for their email (when the identity provider vouched for it), the access-log rows of their own visits, redirects from their old addresses to sites they handed on (those addresses answer not found), and the account. Audit rows stay and keep the opaque user id, and some keep their username or email in the detail, until retention prunes them; the `user_erased` row names them by id only. The address label stays held, and their sign-in identity is held as a SHA-256 of issuer and subject and of their verified email, so they cannot sign in again until an admin allows it (`/api/admin/erased-identities`). Send `confirm` (form or JSON) equal to the username.
      tags: [Admin]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: string
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: string
      responses:
        '200':
          description: Erased
        '400':
          description: The confirmation does not match the username, or the name is a team
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The person can still sign in, or is the last member of a team
  /api/admin/erased-identities:
    get:
      summary: List erased people still refused at sign-in
      description: Admin only, browser session only. One entry per erased person, by the first 12 hex characters of the held hash, with when and by which admin; never a name or address.
      tags: [Admin]
      responses:
        '200':
          description: Held identities, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  erased_identities:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        hash_prefix:
                          type: string
                        erased_at:
                          type: string
                          format: date-time
                        erased_by:
                          type: string
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/erased-identities/{id}/allow:
    post:
      summary: Let an erased person sign in again
      description: Admin only, browser session only. Deletes the held identity; the person's next sign-in makes a new, empty account. Audited as `erased_identity_allowed`.
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Allowed
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No held identity has that id
components:
  headers:
    LatestSkillVersion:
      description: Latest skill bundle version published by this server.
      schema:
        type: string
        example: 0.8.1
    MinimumSkillVersion:
      description: Oldest skill version accepted for API-key management calls.
      schema:
        type: string
        example: 0.8.0
    SkillUpdateURL:
      description: Same-origin page describing the permissioned update flow.
      schema:
        type: string
        example: /install.html
...
    NoStore:
      description: Prevents storage of search responses and telemetry outcomes.
      schema:
        type: string
        example: no-store
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        An API key, for CI and other automation; agents normally connect
        through `/mcp` with the company sign-in instead. Each key has a scope
        chosen when it is created and an expiry (at most 365 days).
        `publish` (the default) deploys, updates and rolls back sites, reads
        versions, saved data and assets, and calls `/api/me` and `/mcp`.
        `full` adds everything else the person can do (deleting sites, access
        levels, viewers, teams, the audit and access logs, search), except
        administration. `offboard` (admins only) calls
        `POST /api/admin/users/disable` and nothing else. No key reaches any
        other `/api/admin/*` route. OAuth connector tokens work only on
        `/mcp`.
  parameters:
    SkillVersion:
      name: X-Skill-Version
      in: header
      required: false
      description: Required on API-key management requests made by an agent. Strict numeric major.minor.patch; the server rejects malformed or below-minimum values.
      schema:
        type: string
        pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$'
      example: 0.8.1
    SimpleHostClient:
      name: X-Simple-Host-Client
      in: header
      required: false
      description: Intentional non-skill client classification. Direct API integrations send `api`; `control-ui` is reserved for the first-party account page. This is compatibility metadata, not authentication, and never replaces X-API-Key.
      schema:
        type: string
        enum: [api, control-ui]
      example: api
...
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        active_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        url:
          type: string
          description: The site's address. Use it as given.
          example: https://my-cool-site.alice-smith.simple-host.example.com/
        new_version:
          type: integer
          description: Only after a deploy with `publish=false`, the version it stored, which is not live.
        analytics:
          $ref: '#/components/schemas/SiteAnalytics'
...
    SiteAnalytics:
      type: object
      description: Server-side daily aggregate analytics for this site.
      properties:
        today_pageviews:
          type: integer
          format: int64
          description: Successful HTML pageviews counted today in UTC.
        today_visits:
          type: integer
          format: int64
          description: Site-scoped 30-minute visits counted today in UTC.
        last_7_pageviews:
          type: integer
          format: int64
          description: Successful HTML pageviews counted over the last 7 UTC days, including today.
        last_7_visits:
          type: integer
          format: int64
          description: Site-scoped visits counted over the last 7 UTC days, including today.
...
    SkillVersionRequired:
      type: object
      required: [error, code, latest_version, minimum_supported_version, update_url]
      properties:
        error:
          type: string
          example: a supported Simple Host skill version is required
        code:
          type: string
          enum: [skill_version_required]
        latest_version:
          type: string
          example: 0.8.2
        minimum_supported_version:
          type: string
          example: 0.8.0
        update_url:
          type: string
          enum: [/install.html]
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        reason:
          type: string
          description: With `site_restricted_by_admin`, the admin's note.
...
  responses:
...
    SkillVersionRequired:
      description: A valid API key was supplied, but the management client omitted a supported skill version or explicit non-skill classification.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Simple-Host-Latest-Skill-Version:
          $ref: '#/components/headers/LatestSkillVersion'
        X-Simple-Host-Minimum-Skill-Version:
          $ref: '#/components/headers/MinimumSkillVersion'
        X-Simple-Host-Update-URL:
          $ref: '#/components/headers/SkillUpdateURL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SkillVersionRequired'
...
    Unauthorized:
      description: 'Missing or invalid credentials. For an `X-API-Key` that did not authenticate, `code` says why: `key_expired` (the body names the date; `expired_at` carries it), `key_revoked`, `key_owner_disabled` or `key_not_recognised`. A key with 14 days or less left gets `X-Key-Expires` (RFC 3339) and `X-Simple-Host-Notice` on every response.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated caller does not have the owner-only permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Site, retained version, or membership was not found or is not accessible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...